Changes from BIGINT branch merged into the trunk - dbid updated from int to bigint - This had a ripple effect, with changes made to all the relevent gratia code that got affected by the bigint change     - Data in all java classes, hibernate configuration files and SQL procedures changed accordingly     - Upgrade86.java code updates the database from version 85 (int dbid) to version 86 (bigint dbid) and all other relevant tables/fields
git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@4120 c85a5f45-a5f3-44fc-b02c-7891c381c519

+++ collector/gratia-services/net/sf/gratia/services/DataScrubber.java
@@ -96,14 +96,14 @@ public class DataScrubber {
             query.setString( "dateLimit", limit );
             
             Logging.debug("DataScrubber: About to execute " + query.getQueryString());
-            Integer result = (Integer) query.uniqueResult();
+            Long result = (Long) query.uniqueResult();
             
             if ( result == null ) {
                query = session.createSQLQuery("select max(dbid) from "+type+"_Meta where ServerDate < :dateLimit");
                query.setString( "dateLimit", limit );
                
                Logging.debug("DataScrubber: About to execute " + query.getQueryString());
-               result = (Integer) query.uniqueResult();
+               result = (Long) query.uniqueResult();
             }
             if ( result == null ) {
                
@@ -120,7 +120,7 @@ public class DataScrubber {
             query = session.createSQLQuery("select dbid from "+type+"_Xml X where ExtraXml = \"\" order by dbid limit 1");
  
             Logging.debug("DataScrubber: About to execute " + query.getQueryString());
-            result = (Integer) query.uniqueResult();
+            result = (Long) query.uniqueResult();
             if ( result == null ) {
                Logging.debug("DataScrubber: found no min(dbid) in "+type+"_Xml");
                // Nothing to do.
